+++ CHANGELOG.md
@@ -5,11 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Custom logos support in dark mode
 ### Changed
 - AdServers' name on flow chart label
 ### Fixed
 - Hide dashboard charts for moderator
 - User list actions for moderator
+- Error notification for Bad Request
 
 ## [1.3.0] - 2023-05-15
 ### Added