+++ CHANGELOG.md
@@ -6,10 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-- Configure creative placeholders
+- Custom logos support in dark mode
+- Set creative placeholders
+### Changed
+- AdServers' name on flow chart label
 ### Fixed
 - Hide dashboard charts for moderator
 - User list actions for moderator
+- Error notification for Bad Request
 
 ## [1.3.0] - 2023-05-15
 ### Added