+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Hide dashboard charts for moderator
 - User list actions for moderator
 - Error notification for Bad Request
+- Warning about refund applies to advertiser commission
 
 ## [1.3.0] - 2023-05-15
 ### Added